Refactor 1
A Character seems like a concept that could exist independent of a
fight, unlike what Character was, so I changed Character to
FightCharacter. And then FightEvent for consistency. I cleaned up
DamageMeter a bit, simplying it a bit and defaulting to "You" if we
couldn't find the owning character name.