Make pet recognition a little more lenient
Problem was we were asserting the owner had to be a PC, but when
PC-knowledge comes in slowly now instead of being awaited, we wouldn't
get an opportunity to recognize as a pet during test/debugging scenarios
w/ static log files. No real harm in opening it up some more, because
the apostrophe-based convention effectively limits false positives
already, regardless of the PC/NPC-ness of the potential owners.